\name{defineMirhostgenes}
\alias{defineMirhostgenes}
\alias{downloadMirbase}
\alias{fetchAdditionalInformation}
\alias{getArrayFeaturesForTx}
\alias{makeHostgeneSQLiteFromTables}
\alias{makeMirhostgenesPackage}
%\alias{createPremirnaSequenceTable}
%\alias{createMirfamTable}

\title{
  Utilities to define miRNA host genes and build a corresponding
  annotation package.
}
\description{
  These functions allow download a specific miRBase release, define
  miRNA host genes for miRNAs of a species, generate an SQLite database
  containing that information and ultimately build the corresponding
  annotation package inR.
}
\usage{

%createMirfamTable(x, premirna.file="pre_mirna.txt")

%createPremirnaSequenceTable(x, premirna.file="pre_mirna.txt")

defineMirhostgenes(gff, database=c("core"), host="ensembldb.ensembl.org",
                   user="anonymous", pass, ensemblapi, verbose=FALSE)

downloadMirbase(version, path=".", force.download=FALSE)

fetchAdditionalInformation(mirbase.path=".", path=".", verbose=FALSE)

getArrayFeaturesForTx(species, arrays=c("HG-U133_Plus_2", "PrimeView"),
                      prop.probes=0.8, max.mm=0, min.probe.algn=24,
                      host="ensembldb.ensembl.org", user="anonymous",
                      pass, ensemblapi, verbose=FALSE)

makeHostgeneSQLiteFromTables(path=".")

makeMirhostgenesPackage(db, version, maintainer, author, destDir=".",
                        license="Artistic-2.0")


}
%- maybe also 'usage' for other objects documented here.
\arguments{

  \item{arrays}{
    Character vector specifying the types of microarrays for which probe
    sets should be searched for. Note that these have to correspond to
    the names of microarrays as available in the Ensembl databases.
  }

  \item{author}{
    The author of the package.
  }

  \item{database}{
    A character vector with the database name(s) that should be
    queried. Allowed are \code{"core"}, \code{"cdna"},
    \code{"otherfeatures"} and \code{"vega"}. For most cases it suffices
    to query the Ensembl \code{"core"} database, as gene and transcript
    models are mostly redundant between the databases (with \code{"vega"}
    containing human curated genes and \code{"otherfeatures"} NCBI
    RefSeq genes and other features).
  }

  \item{db}{
    For \code{makeMirhostgenesPackage}: the file name of the (SQLite)
    database file created with the function
    \code{makeHostgeneSQLiteFromTables}.

  }

  \item{destDir}{
    Where the package should be saved to.
  }

  \item{gff}{
    The gff file name containing the genomic alignments for the
    pre-miRNAs. Such gff files (one per species) are located in the
    genomes folder of the downloaded miRBase resource (e.g. downloaded
    by the \code{downloadMirbase} function). Note that the specified gff
    has to be in gff version 3 format, thus a \code{".gff3"}
    (e.g. \code{"hsa.gff3"}) file from the genomes forldes shoudl be
    submitted.
  }

  \item{ensemblapi}{
    The path to the Ensembl perl API installed locally on the
    system. The Ensembl perl API version determines which Ensembl
    database version is queried.
  }

  \item{force.download}{
    Force the download of the miRBase even if the same version is
    already available locally.
  }

  \item{host}{
    The hostname to access the Ensembl database.
  }

  \item{license}{
    The license of the package.
  }

  \item{maintainer}{
    The maintainer of the package.
  }

  \item{max.mm}{
    Maximum number of mismatches of a probe with the target genes.
  }

  \item{min.probe.algn}{
    Minimal length of the probe alignment within the exons of a
    transcript. The default value of 24 means that all nucleotides of a
    25nt long probe have to map within the exons of a transcript.
  }

  \item{mirbase.path}{
    For \code{fetchAdditionalInformation}: the directory to which the
    miRBase database files have been downloaded (by the
    \code{downloadMirbase} function).
  }

  \item{pass}{
    The password for the Ensembl database.
  }

  \item{path}{
    \code{makeHostgenesSQLiteFromTables}: the directory in which the
    database table files generated by \code{defineHostgenes}
    The directory to which \code{downloadMirbase} will download the
    miRBase and in which \code{makeHostgeneSQLiteFromTables} can find
    the txt files generated by \code{defineMirhostgenes} are located.
  }

  %% \item{premirna.file}{
  %%   For function \code{createPremirnaSequenceTable}: the tabulator
  %%   delimited text file containing the miRBase pre-miRNA IDs. This txt
  %%   file is created by the \code{defineMirhostgenes} function.
  %% }

  \item{prop.probes}{
    Proportion of probes of a probe set that have to map within the
    exons of a transcript to be considered. The default value of 0.8
    means that at least 80 percent of the probes of a probe set have to
    target the transcript.
  }

  \item{species}{
    For \code{getArrayFeaturesForTx}: the species the miRNA host genes were defined for.
  }

  \item{user}{
    The username for the Ensembl database.
  }

  \item{verbose}{
    print progress messages.
  }

  \item{version}{
    For \code{downloadMirbase}: the version of the miRBase that should
    be downloaded. If not specified the most recent version is
    downloaded.

    For \code{makeMirhostgenesPackage}: the version for the package to
    be created.
  }

  %% \item{x}{
  %%   For \code{createPremirnaSequenceTable}: the fasta file containing
  %%   the sequences for the pre-miRNAs. This file (\code{"hairpin.fa.zip"}) is included in the
  %%   miRBase download and is by default also downloaded by the function
  %%   \code{downloadMirbase}. Both the zipped or unzipped file can be
  %%   submitted to the function.

  %%   For \code{createMirfamTable}: the file from the miRBase providing
  %%   the miRNA family information (usually called \code{"miFam.dat"}). This
  %%   file is downloaded along with all other miRBase ressources by the
  %%   \code{downloadMirbase} function.
  %% }


}
\details{

  The \code{downloadMirbase} and \code{defineMirhostgenes} functions
  internally call the perl scripts \code{get-mirbase.pl} and
  \code{define_mirna_host_genes.pl}, respectively. The
  \code{define_mirna_host_genes.pl} needs the Perl API and bioperl to be
  present in the \code{PERL5LIB} environment variable.

  %% The \code{createMirfamTable} function reads the miRNA family
  %% definition provided in the \code{"miFam.dat"} file from the miRBase
  %% and saves the miRNA family information for the pre-miRNAs for which
  %% host genes were defined by the \code{defineMirhostgenes} to a text
  %% file \code{"mirfam.txt"} in the current R working directory. This
  %% function should be called after \code{defineMirhostgenes}, but
  %% before \code{makeHostgeneSQLiteFromTables} (see examples below).

  The \code{fetchAdditionalInformation} function extracts additional
  informations such as the confidence information, read counts,
  pre-miRNA sequences and miRNA family definitions from the downloaded
  miRBase database tables and inserts it into tables for the
  \code{MirhostDb} database. This function should be called after
  \code{defineMirhostgenes} and before \code{makeHostgeneSQLiteFromTables}.

  %% The \code{createPremirnaSequenceTable} function reads the fasta file
  %% from miRBase and writes the sequences of all pre-miRNAs specified in
  %% the \code{premirna.file} file to a table with the name \code{"pre_mirna_sequence.txt"} to
  %% the current directory. This function should be called after
  %% \code{defineMirhostgenes}, but before
  %% \code{makeHostgeneSQLiteFromTables} (see examples below).

  The \code{getArrayFeaturesForTx} again uses the Ensembl Perl API to
  fetch, for the defined host transcripts, microarray features (probe
  sets) possibly detecting the transcripts.

  The \code{makeHostgeneSQLiteFromTables} function reads all the txt
  files generated by the \code{defineMirhostgenes} function and builds
  a SQLite database. If additional files \code{"pre_mirna_sequence.txt"}
  and/or \code{"mirna_fam.txt"}, created by the functions
  \code{createPremirnaSequenceTable} and \code{createMirfamTable} are also
  present, the information contained in these files will be added to the
  database too.

  The \code{makeMirhostgenesPackage} finally creates an annotation
  package based on the SQLite file generated above.

}
\note{
  A local installation of the Ensembl perl API is required for the
  \code{defineMirhostgenes}. See
  \url{http://www.ensembl.org/info/docs/api/api_installation.html} for
  installation inscructions.
}
\author{
  Johannes Rainer
}
\examples{

\dontrun{

library(mirhostgenes)

## Download mirbase version 20 (matching genome release 37)
downloadMirbase(version=20)

## Define miRNA host genes using the Ensembl core database.
## we're using the gff file for human miRNAs of the miRBase version we
## just downloaded.
## we set v=TRUE to get some feedback about the progress.
defineMirhostgenes(gff="20/genomes/hsa.gff3", v=TRUE)

## Fetch additional information from downloaded miRBase files:
## o pre-miRNA sequence data.
## o miRNA family information.
## o pre- and mature miRNA confidence data.
## o pre- and mature miRNA read count data.
fetchAdditionalInformation(mirbase.path="20/")

## Add probe features... for Affymetrix microarrays. It is crucial that
## the species matches!
## We do also specify form which microarrays we want to fetch the probes/
## probe sets.
getArrayFeaturesForTx(species="human", arrays=c("HG-U133_Plus_2", "PrimeView"))

## Build the SQLite database from the generated txt files.
DBNAME <- makeHostgeneSQLiteFromTables()

## Build a R package providing the annotation database.
makeMirhostgenesPackage(DBNAME,
                        version="0.0.1",
                        maintainer="Johannes Rainer <johannes.rainer@eurac.edu>",
                        author="J Rainer"
                        )

}

}
\keyword{ data }

