\name{PositionFilter-class}
\Rdversion{1.1}
\docType{class}
\alias{AlignmentidFilter-class}
\alias{ArrayFilter-class}
\alias{DatabaseFilter-class}
\alias{PositionFilter-class}
\alias{PremirnaConfidence-class}
\alias{PremirnaFilter-class}
\alias{PremirnaidFilter-class}
\alias{ProbesetidFilter-class}
\alias{MatmirnaConfidence-class}
\alias{MatmirnaFilter-class}
\alias{MatmirnaidFilter-class}
\alias{MirfamFilter-class}
\alias{MirfamidFilter-class}
\alias{ReadCountFilter-class}
% column, Filter, MirhostDb, missing
\alias{column,AlignmentidFilter,MirhostDb,missing-method}
\alias{column,ArrayFilter,MirhostDb,missing-method}
\alias{column,DatabaseFilter,MirhostDb,missing-method}
\alias{column,MatmirnaConfidence,MirhostDb,missing-method}
\alias{column,MatmirnaFilter,MirhostDb,missing-method}
\alias{column,MatmirnaidFilter,MirhostDb,missing-method}
\alias{column,MirfamFilter,MirhostDb,missing-method}
\alias{column,MirfamidFilter,MirhostDb,missing-method}
\alias{column,PositionFilter,MirhostDb,missing-method}
\alias{column,PremirnaConfidence,MirhostDb,missing-method}
\alias{column,PremirnaFilter,MirhostDb,missing-method}
\alias{column,PremirnaidFilter,MirhostDb,missing-method}
\alias{column,ProbesetidFilter,MirhostDb,missing-method}
\alias{column,ReadCountFilter,MirhostDb,missing-method}
% where, Filter, MirhostDb, missing
\alias{where,AlignmentidFilter,MirhostDb,missing-method}
\alias{where,ArrayFilter,MirhostDb,missing-method}
\alias{where,DatabaseFilter,MirhostDb,missing-method}
\alias{where,MatmirnaConfidence,MirhostDb,missing-method}
\alias{where,MatmirnaFilter,MirhostDb,missing-method}
\alias{where,MatmirnaidFilter,MirhostDb,missing-method}
\alias{where,MirfamFilter,MirhostDb,missing-method}
\alias{where,MirfamidFilter,MirhostDb,missing-method}
\alias{where,PositionFilter,MirhostDb,missing-method}
\alias{where,PremirnaConfidence,MirhostDb,missing-method}
\alias{where,PremirnaFilter,MirhostDb,missing-method}
\alias{where,PremirnaidFilter,MirhostDb,missing-method}
\alias{where,ProbesetidFilter,MirhostDb,missing-method}
\alias{where,ReadCountFilter,MirhostDb,missing-method}
% column, Filter, MirhostDb, character
\alias{column,AlignmentidFilter,MirhostDb,character-method}
\alias{column,ArrayFilter,MirhostDb,character-method}
\alias{column,DatabaseFilter,MirhostDb,character-method}
\alias{column,MatmirnaConfidence,MirhostDb,character-method}
\alias{column,MatmirnaFilter,MirhostDb,character-method}
\alias{column,MatmirnaidFilter,MirhostDb,character-method}
\alias{column,MirfamFilter,MirhostDb,character-method}
\alias{column,MirfamidFilter,MirhostDb,character-method}
\alias{column,PositionFilter,MirhostDb,character-method}
\alias{column,PremirnaConfidence,MirhostDb,character-method}
\alias{column,PremirnaFilter,MirhostDb,character-method}
\alias{column,PremirnaidFilter,MirhostDb,character-method}
\alias{column,ProbesetidFilter,MirhostDb,character-method}
\alias{column,ReadCountFilter,MirhostDb,character-method}
% where, Filter, MirhostDb, character
\alias{where,AlignmentidFilter,MirhostDb,character-method}
\alias{where,ArrayFilter,MirhostDb,character-method}
\alias{where,DatabaseFilter,MirhostDb,character-method}
\alias{where,MatmirnaConfidence,MirhostDb,character-method}
\alias{where,MatmirnaFilter,MirhostDb,character-method}
\alias{where,MatmirnaidFilter,MirhostDb,character-method}
\alias{where,MirfamFilter,MirhostDb,character-method}
\alias{where,MirfamidFilter,MirhostDb,character-method}
\alias{where,PositionFilter,MirhostDb,character-method}
\alias{where,PremirnaConfidence,MirhostDb,character-method}
\alias{where,PremirnaFilter,MirhostDb,character-method}
\alias{where,PremirnaidFilter,MirhostDb,character-method}
\alias{where,ProbesetidFilter,MirhostDb,character-method}
\alias{where,ReadCountFilter,MirhostDb,character-method}
% column
\alias{where,list,MirhostDb,missing,missing-method}
\alias{column,AlignmentidFilter,missing,missing-method}
\alias{column,ArrayFilter,missing,missing-method}
\alias{column,DatabaseFilter,missing,missing-method}
\alias{column,MatmirnaConfidence,missing,missing-method}
\alias{column,MatmirnaFilter,missing,missing-method}
\alias{column,MatmirnaidFilter,missing,missing-method}
\alias{column,MirfamFilter,missing,missing-method}
\alias{column,MirfamidFilter,missing,missing-method}
\alias{column,PositionFilter,missing,missing-method}
\alias{column,PremirnaConfidence,missing,missing-method}
\alias{column,PremirnaFilter,missing,missing-method}
\alias{column,PremirnaidFilter,missing,missing-method}
\alias{column,ProbesetidFilter,missing,missing-method}
\alias{column,ReadCountFilter,missing,missing-method}
% where
\alias{where,AlignmentidFilter,missing,missing-method}
\alias{where,ArrayFilter,missing,missing-method}
\alias{where,DatabaseFilter,missing,missing-method}
\alias{where,MatmirnaConfidence,missing,missing-method}
\alias{where,MatmirnaFilter,missing,missing-method}
\alias{where,MatmirnaidFilter,missing,missing-method}
\alias{where,MirfamFilter,missing,missing-method}
\alias{where,MirfamidFilter,missing,missing-method}
\alias{where,PositionFilter,missing,missing-method}
\alias{where,PremirnaConfidence,missing,missing-method}
\alias{where,PremirnaFilter,missing,missing-method}
\alias{where,PremirnaidFilter,missing,missing-method}
\alias{where,ProbesetidFilter,missing,missing-method}
\alias{where,ReadCountFilter,missing,missing-method}
\alias{where,list,MirhostDb,ANY-method}
% value
\alias{value,AlignmentidFilter-method}
\alias{value,ArrayFilter-method}
\alias{value,DatabaseFilter-method}
\alias{value,MatmirnaConfidence-method}
\alias{value,MatmirnaFilter-method}
\alias{value,MatmirnaidFilter-method}
\alias{value,MirfamFilter-method}
\alias{value,MirfamidFilter-method}
\alias{value,PositionFilter-method}
\alias{value,PremirnaConfidence-method}
\alias{value,PremirnaFilter-method}
\alias{value,PremirnaidFilter-method}
\alias{value,ProbesetidFilter-method}
\alias{value,ReadCountFilter-method}
% value<-
\alias{value<-,AlignmentidFilter-method}
\alias{value<-,ArrayFilter-method}
\alias{value<-,DatabaseFilter-method}
\alias{value<-,MatmirnaConfidence-method}
\alias{value<-,MatmirnaFilter-method}
\alias{value<-,MatmirnaidFilter-method}
\alias{value<-,MirfamFilter-method}
\alias{value<-,MirfamidFilter-method}
\alias{value<-,PositionFilter-method}
\alias{value<-,PremirnaConfidence-method}
\alias{value<-,PremirnaFilter-method}
\alias{value<-,PremirnaidFilter-method}
\alias{value<-,ProbesetidFilter-method}
\alias{value<-,ReadCountFilter-method}
% condition
\alias{condition,AlignmentidFilter-method}
\alias{condition,ArrayFilter-method}
\alias{condition,DatabaseFilter-method}
\alias{condition,MatmirnaConfidence-method}
\alias{condition,MatmirnaFilter-method}
\alias{condition,MatmirnaidFilter-method}
\alias{condition,MirfamFilter-method}
\alias{condition,MirfamidFilter-method}
\alias{condition,PositionFilter-method}
\alias{condition,PremirnaConfidence-method}
\alias{condition,PremirnaFilter-method}
\alias{condition,PremirnaidFilter-method}
\alias{condition,ProbesetidFilter-method}
\alias{condition,ReadCountFilter-method}
% condition<-
\alias{condition<-,AlignmentidFilter-method}
\alias{condition<-,ArrayFilter-method}
\alias{condition<-,DatabaseFilter-method}
\alias{condition<-,MatmirnaConfidence-method}
\alias{condition<-,MatmirnaFilter-method}
\alias{condition<-,MatmirnaidFilter-method}
\alias{condition<-,MirfamFilter-method}
\alias{condition<-,MirfamidFilter-method}
\alias{condition<-,PositionFilter-method}
\alias{condition<-,PremirnaConfidence-method}
\alias{condition<-,PremirnaFilter-method}
\alias{condition<-,PremirnaidFilter-method}
\alias{condition<-,ProbesetidFilter-method}
\alias{condition<-,ReadCountFilter-method}
\alias{column}
\alias{where}

\title{Filter results fetched from the miRNA host genes database}

\description{
  These classes allow to specify which entries (i.e. mature miRNAs,
  pre-miRNAs etc. should be retrieved from the database).
}

\section{Objects from the Class}{
  While objects can be created by calls
  e.g. of the form \code{new("PositionFilter", ...)} users are strongly
  encouraged to use the specific functions:
  \code{\link{AlignmentidFilter}}, \code{\link{ArrayFilter}},
  \code{\link{DatabaseFilter}}, \code{\link{MatmirnaConfidence}},
  \code{\link{MatmirnaFilter}}, \code{\link{MatmirnaidFilter}},
  \code{\link{MirfamFilter}} and \code{\link{MirfamidFilter}},
  \code{\link{PositionFilter}}, \code{\link{PremirnaConfidence}}
  \code{\link{PremirnaFilter}}, \code{\link{PremirnaidFilter}},
  \code{\link{ProbesetidFilter}} and \code{\link{ReadCountFilter}}.  See
  examples below for usage.
}

\section{Slots}{
  \describe{
    \item{condition}{
      Object of class \code{"character"}: can be
      either \code{"="}, \code{"in"} or \code{"like"} to filter on character values
      (e.g. gene id, gene biotype, seqname etc), or \code{"="}, \code{">"}
      or \code{"<"} for numerical values (chromosome/seq
      coordinates). Note that for \code{"like"} \code{value} should be a
      SQL pattern (e.g. \code{"ENS\%"} that matches all values starting
      with \code{"ENS"}).
    }

    \item{match.case}{
      Object of class \code{"logical"}: for
      \code{PremirnaFilter} and \code{MatmirnaFilter} specifying whether
      case sensitive or in-sensitive queries should be performed for
      the pre-miRNA or mature miRNA names.
    }

    \item{of}{
      For \code{ReadCountFilter}: spefifying the database table in which
      the read counts are stored (i.e. either \code{"pre_mirna"} or
      \code{"mat_mirna"}).
    }

    \item{value}{
      Object of class \code{"character"}: the value
      to be used for filtering.
    }

  }
}
\section{Extends}{
  Class \code{"BasicFilter"} defined in the \code{ensembldb} package, directly.
}
\section{Methods}{

  (in alphabetic order)

  \describe{

    \item{column}{
      \code{signature(object = "PositionFilter", db = "MirhostDb",
      with.tables="character")}: returns the column (table column name
      in the miRNA host genes database) to be used for the
      filtering. The method can be applied to any filter class defined
      in the \code{mirhostgenes} or \code{ensembldb} packages. Argument
      \code{with.tables} allows to restrict the filter to be applied to
      certain database tables. By default the filter is applied to all
      database tables.

      The method returns a character string with the name of the column.
    }

    \item{condition, condition<-}{
      Get and set the \code{condition} of the filter (i.e. how values
      should be compared).

      The \code{condition} method returns a character string
      representing the condition.
    }

    \item{value, value<-}{
      Get and set the \code{value} of the filter.

      The \code{value} method returns a character vector of the
      value(s).
    }

    \item{where}{
      \code{signature(object = "PositionFilter", db = "MirhostDb",
      with.tables="character")}: returns the where clause for the SQL
      call to the miRNA host genes database. The method can be applied
      to any filter class defined in the \code{mirhostgenes} or
      \code{ensembldb} packages. Argument
      \code{with.tables} allows to restrict the filter to be applied to
      certain database tables. By default the filter is applied to all
      database tables.

      The method returns the SQL \emph{where} clause that will be used
      to restrict the results returned by the database. This string is
      dependent on the type of database.
    }

  }
}
\details{
  Description of filters (in alphabetical order).

  \describe{
    \item{AlignmentidFilter}{
      Allows to filter results based on the (internal) alignment ID
      for a pre-miRNA. The same pre-miRNA sequence might have multiple
      genomic alignments, thus, the alignment ID was introduced
      ensure an unique identifier in the database. Both, the mature
      miRNAs and the host transcripts are linked via this ID
      to the pre-miRNA alignment.
    }

    \item{ArrayFilter}{
      Allows to filter results based on provided
      microarray names (e.g. \code{"HG-U133_Plus_2"}). Use the method
      \code{\link{listArrays}} to get a list of available microarray names.
    }

    \item{DatabaseFilter}{
      Allows to restrict results to genes/transcripts defined in the
      speciefied databases. Since the database contains gene/transcripts
      defined in various databases (e.g. Ensembl, Vega, NCBI RefSeq)
      restricting the results to one database helps avoiding redundant
      gene/transcript definitions.
      Use the method \code{\link{listDatabases}}
      to get a list of supported values.
    }

    \item{MarmirnaConfidence}{
      Allows to filter on the confidence information for mature and
      pre-miRNAs stored in the miRBase. Requires that the corresponding
      information have been extracted from the miRBase and inserted into
      the \code{MirhostDb} database. Note that miRBase prior version 21 did
      not provide mature miRNA confidence information.
      Also, at present, the filter allows only to filter the query to
      fetch only high confidence (by specifying \code{value="high"})
      miRNAs.
    }

    \item{MatmirnaFilter}{
      Filter on the mature miRNA names (e.g. \code{"hsa-miR-15b-5p"})
    }

    \item{MatmirnaidFilter}{
      Filter on the mature miRNA identifiers (i.e. mirbase accession
      numbers, e.g. \code{"MIMAT0000062"}).
    }

    \item{PositionFilter}{
      Allows to filter results based on whether the (mature)
      miRNA is exonic or intronic, i.e. is encoded in an exon
      of the transcript, or in an intron.
    }

    \item{PremirnaConfidence}{
      See \code{MatmirnaConfidence} above for a description. In contrast
      to mature miRNAs, also miRBase versions prior to version 21 did
      provide high confidence information.
    }

    \item{PremirnaFilter}{
      Allows to filter on pre-miRNA names (e.g. \code{"hsa-mir-15b"}).
    }

    \item{PremirnaidFilter}{
      Allows to filter on pre-miRNA identifiers (mirbase accession
      numbers, e.g. \code{"MI0000060"}).
    }

    \item{ProbesetidFilter}{
      Allows to filter on microarray probe set IDs.
    }

    \item{ReadCountFilter}{
      Allows to filter the results by the number of reads from various
      high throughput sequencing experiments mapping to the mature or
      pre-miRNAs.
    }

    In addition to the above classes all Filter classes defined in the
    \code{ensembldb} package. See the vignette for a more detailed
    description.
  }

}
\note{
  All methods for filter classes (also those defined in the
  \code{ensembldb} package,
  i.e. \code{\link[ensembldb]{GeneidFilter}},
  \code{\link[ensembldb]{GenenameFilter}},
  \code{\link[ensembldb]{GenebiotypeFilter}},
  \code{\link[ensembldb]{TxidFilter}},
  \code{\link[ensembldb]{TxbiotypeFilter}},
  \code{\link[ensembldb]{SeqnameFilter}},
  \code{\link[ensembldb]{SeqstrandFilter}},
  \code{\link[ensembldb]{SeqstartFilter}},
  \code{\link[ensembldb]{SeqendFilter}}) should include the
  \code{MirhostDb} class instance as this will cause the
  methods (\code{column} and \code{where}) to return values
  specific for the layout of the miRNA host genes database. See
  examples below for more details.
}
\value{
  See method descriptions above for the return values.
}

\author{
  Johannes Rainer
}
\seealso{
  \code{\link[ensembldb]{GeneidFilter}},
  \code{\link[ensembldb]{SeqnameFilter}},
  \code{\link{GeneidFilter}} (in the \code{mirhostgenes} package).
}
\examples{

library(MirhostDb.Hsapiens.v75.v20)
mhdb <- MirhostDb.Hsapiens.v75.v20

## Create a filter that could be used to retrieve all informations for
## the respective gene.
Gif <- GeneidFilter("ENSG00000012817")
Gif
## Returns the where clause of the SQL querys specific for the
## mirhostgenes database
where(Gif, mhdb)

## Create a filter for a chromosomal end position of a gene
Sef <- SeqendFilter(10000, condition=">", "gene")
Sef

## Get and set the value of a filter.
value(Sef)
value(Sef) <- 100001

## In the miRNA host genes database we are rather interested to get
## pre-miRNAs from certain genomic positions, thus we define "pre_mirna"
## as the feature and call where by supplying also the MirhostDb
## object
Sef <- SeqendFilter(1000, condition=">", "pre_mirna")
where(Sef, mhdb)

## Get all high confidence pre-miRNAs.
premirnas(mhdb, filter=PremirnaConfidence("high"))

## Get all mature miRNAs with more than 1000 reads across all experiments
matmirnas(mhdb, filter=ReadCountFilter(1000, of="mat_mirna"))

}
\keyword{classes}

